Add files via upload
This is the first script to run, it extracts the Entrez Gene names from a gff3 file. This script is needed only if your two alignments have different ensembl gene identifiers. In this instance, the annotation file had D2 specific ensemble IDs which needed to be converted into B6 specific ensembl IDs.